Rename UserLocationVO class
